[CMakeLists.txt] Update core dependencies to conditionally include OpenGL and glad based on UI backend selection [ApplicationComputeRuntime.cpp] Add WebGPU context retrieval method and store context in member variable [ApplicationComputeRuntime.h] Declare virtual method for retrieving WebGPU context in compute runtime interface [CMakeLists.txt] Adjust source file inclusion based on selected UI backend for OpenGL and WebGPU [FunctionFromImage3DView.cpp] Wrap OpenGL specific code with preprocessor directives for backend compatibility [GLView.cpp] Implement backend-specific initialization and shutdown for OpenGL and WebGPU [GLView.h] Add member variables and methods for managing WebGPU context and rendering [ImageStackView.cpp] Modify texture upload logic to accommodate backend-specific implementations [MainWindow.cpp] Set WebGPU context in main view upon runtime initialization [RenderWindow.cpp] Include WebGPU frame presenter for rendering [RenderWindow.h] Update frame presenter to be backend-neutral [RenderWindowWebGPU.cpp] Implement WebGPU frame presenter logic for rendering [ThreemfThumbnailExtractor.cpp] Adjust thumbnail texture handling for backend compatibility [ThreemfThumbnailExtractor.h] Change thumbnail texture ID type to backend-neutral handle [FramePresenter.h] Introduce backend-neutral frame presenter interface for rendering [OpenGLFramePresenter.cpp] Update texture ID return type to uintptr_t for consistency [OpenGLFramePresenter.h] Inherit from FramePresenter and implement required methods [WebGPUFramePresenter.cpp] Implement WebGPU frame presenter for handling texture uploads [WebGPUFramePresenter.h] Define WebGPU frame presenter interface for backend-neutral rendering [WebGPUComputeContext.cpp] Add methods to retrieve WebGPU instance and adapter [WebGPUComputeContext.h] Declare methods for accessing WebGPU instance and adapter [WebGPUComputeRenderer.cpp] Correctly set height in render frame based on viewport

+++ gladius/src/ui/render/OpenGLFramePresenter.cpp
@@ -63,7 +63,7 @@ namespace gladius::ui::async_rendering
         m_freshness.reset();
     }
 
-    std::uint32_t OpenGLFramePresenter::getTextureId() const noexcept
+    std::uintptr_t OpenGLFramePresenter::getTextureId() const noexcept
     {
         return m_textureId;
     }
